+++ src/redturtle/prenotazioni/restapi/services/booking/notify_about_confirm.py
@@ -1,11 +1,15 @@
 # -*- coding: utf-8 -*-
 from plone import api
+from plone.protect.interfaces import IDisableCSRFProtection
 from plone.restapi.services import Service
 from Products.DCWorkflow.events import AfterTransitionEvent
 from zExceptions import BadRequest
 from zope.event import notify
+from zope.interface import alsoProvides
 from zope.interface import implementer
 from zope.publisher.interfaces import IPublishTraverse
+
+from redturtle.prenotazioni import _
 
 
 @implementer(IPublishTraverse)
@@ -18,13 +22,20 @@ class NotifyUserAboutBookingConfirm(Service):
         return self
 
     def reply(self):
+        alsoProvides(self.request, IDisableCSRFProtection)
+
         booking = api.content.get(UID=self.booking_uid)
 
         if not booking:
             raise BadRequest("Booking not found")
 
         transition_billet = type(
-            "TranstionBillet", (object,), {"__name__": "confirm", "title": "Confirm"}
+            "TransitionBillet",
+            (object,),
+            {
+                "__name__": "confirm",
+                "title": _("confirm", default="Confirm"),
+            },
         )()
 
         notify(